requirements.txt
docopt
colorama
termcolor
apache-libcloud
pyyaml
